requirements.txt
pytorch-lightning==1.1.1
torchtext==0.8
fairscale==0.4.1
dictdiffer
googletrans
pillow
matplotlib
nltk
torchmetrics
sentencepiece
pandas==1.0.1